Extract microgrid mocking wrapper into a separate module
This can then be used in other tests that need a microgrid mock.
